Cards can now be unselected after time is up.
Before, players would do a turn and the score would be based on the words they have picked.
Now, when the time runs out, players can remove some of the picked words to adjust their score. This helps in case a Team is trying to cheat(selecting words that were not guessed) or if the team tapped on the word by mistake and need to correct it.
To do so, I have added the removeTap function to the  statement on the CardAdapter.java.

Rosário Pereira Fernandes
rosariofernandes51@gmail.com